chore: refactor component structure
