+++ frontend/src/components/header/header.scss
@@ -1,3 +1,6 @@
+@import '../../css/vars';
+@import '../../css/base';
+
 
 .header {
   background: var(--header-background);
