+++ frontend/src/components/toaster/toaster.scss
@@ -1,4 +1,4 @@
-.toasts {
+.toaster {
   list-style: none;
   padding: 0;
   margin: 0;
