+++ frontend/src/css/index.scss
@@ -1,9 +1,6 @@
 @import './vars';
 @import './base';
-@import './header';
 @import './forms';
 @import './login';
-@import './result-list';
 @import './profile-page';
 @import './skill-page';
-@import './toasts';
